SMHasher: fixed 64-bit incompatibility in Platform.h

+++ SMHasher/Platform.h
@@ -75,7 +75,7 @@ __inline__ unsigned long long int rdtsc()
 #ifdef __x86_64__
     unsigned int a, d;
     __asm__ volatile ("rdtsc" : "=a" (a), "=d" (d));
-    return (unsigned long)a | ((unsigned long)d << 32);
+    return (unsigned long)a | ((unsigned long long)d << 32);
 #elif defined(__i386__)
     unsigned long long int x;
     __asm__ volatile ("rdtsc" : "=A" (x));
